Move MockPlatformUtil and MockRequestSource
Part of an effort to attempt to reduce memory usage during
the build, to address ASAN flakes in CI.

- This is a mechanical change, no functional changes.
- Does have some additional cleanup.
