+++ test/mocks.cc
@@ -4,20 +4,14 @@
 
 namespace Nighthawk {
 
-MockPlatformUtil::MockPlatformUtil() = default;
-
 MockBenchmarkClientFactory::MockBenchmarkClientFactory() = default;
 
 MockSequencerFactory::MockSequencerFactory() = default;
-
-MockStoreFactory::MockStoreFactory() = default;
 
 MockStatisticFactory::MockStatisticFactory() = default;
 
 MockRequestSourceFactory::MockRequestSourceFactory() = default;
 
 MockTerminationPredicateFactory::MockTerminationPredicateFactory() = default;
-
-MockRequestSource::MockRequestSource() = default;
 
 } // namespace Nighthawk
