+++ test/mocks.h
@@ -1,40 +1,18 @@
 #pragma once
 
-#include <chrono>
-#include <memory>
-
 #include "envoy/api/api.h"
 #include "envoy/common/time.h"
 #include "envoy/event/dispatcher.h"
-#include "envoy/stats/store.h"
 #include "envoy/upstream/cluster_manager.h"
 
 #include "nighthawk/client/factories.h"
-#include "nighthawk/common/platform_util.h"
-#include "nighthawk/common/request_source.h"
 #include "nighthawk/common/statistic.h"
-#include "nighthawk/common/uri.h"
 
-#include "external/envoy/test/test_common/simulated_time_system.h"
-
-#include "common/utility.h"
-
-#include "absl/types/optional.h"
 #include "gmock/gmock.h"
-
-using namespace std::chrono_literals;
 
 namespace Nighthawk {
 
 // TODO(oschaaf): split this out in files for common/ and client/ mocks
-
-class MockPlatformUtil : public PlatformUtil {
-public:
-  MockPlatformUtil();
-
-  MOCK_CONST_METHOD0(yieldCurrentThread, void());
-  MOCK_CONST_METHOD1(sleep, void(std::chrono::microseconds));
-};
 
 class MockBenchmarkClientFactory : public Client::BenchmarkClientFactory {
 public:
@@ -56,12 +34,6 @@ public:
                                           TerminationPredicatePtr&& termination_predicate,
                                           Envoy::Stats::Scope& scope,
                                           const Envoy::MonotonicTime scheduled_starting_time));
-};
-
-class MockStoreFactory : public Client::StoreFactory {
-public:
-  MockStoreFactory();
-  MOCK_CONST_METHOD0(create, Envoy::Stats::StorePtr());
 };
 
 class MockStatisticFactory : public Client::StatisticFactory {
@@ -87,13 +59,6 @@ public:
                      TerminationPredicatePtr(Envoy::TimeSource& time_source,
                                              Envoy::Stats::Scope& scope,
                                              const Envoy::MonotonicTime scheduled_starting_time));
-};
-
-class MockRequestSource : public RequestSource {
-public:
-  MockRequestSource();
-  MOCK_METHOD0(get, RequestGenerator());
-  MOCK_METHOD0(initOnThread, void());
 };
 
 } // namespace Nighthawk
