+++ test/mocks/common/mock_platform_util.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_platform_util.h"
+
+namespace Nighthawk {
+
+MockPlatformUtil::MockPlatformUtil() = default;
+
+}
