+++ test/mocks/common/mock_request_source.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_request_source.h"
+
+namespace Nighthawk {
+
+MockRequestSource::MockRequestSource() = default;
+
+} // namespace Nighthawk
